% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app_blorr}
\alias{launch_app_blorr}
\title{Launch Logistic Regression App}
\usage{
launch_app_blorr()
}
\description{
Launches the binary logistic regression app
}
\examples{
\dontrun{
launch_app_blorr()
}
}
