% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app_descriptr}
\alias{launch_app_descriptr}
\title{Launch Descriptive Statistics App}
\usage{
launch_app_descriptr()
}
\description{
Launches the descriptr app
}
\examples{
\dontrun{
launch_app_descriptr()
}
}
