% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app_visualizer}
\alias{launch_app_visualizer}
\title{Launch Visualization App}
\usage{
launch_app_visualizer()
}
\description{
Launches the Visualizer app
}
\examples{
\dontrun{
launch_app_visualizer()
}
}
