% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplorerr.R
\docType{package}
\name{xplorerr}
\alias{xplorerr}
\alias{xplorerr-package}
\title{\code{xplorerr} package}
\description{
R Shiny app for interactive statistical analysis
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/xplorerr}{GitHub}
}
